controllers1

+++ src/main/java/com/departement/fichedevoeux/repository/ConversationRepository.java
@@ -6,6 +6,6 @@ import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ConversationRepository extends JpaRepository<Conversation, Long> {
+public interface ConversationRepository extends JpaRepository<Conversation, Integer> {
 	List<Conversation> findByInitiateurId(Long id);
 }
