+++ src/main/java/com/departement/fichedevoeux/repository/ModuleRepository.java
@@ -5,7 +5,7 @@ import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ModuleRepository extends JpaRepository <Module, Long> {
+public interface ModuleRepository extends JpaRepository <Module, Integer> {
 	List <Module> findByNomContainingIgnoreCase(String nom);
 	List<Module> findBySemestre(String semestre);
 	List <Module> findByPallier(String pallier);
